Move all basemap logic to a single hook
